grpc: extract CredentialsBuilders from Otel
