+++ modules/grpc/Makefile.am
@@ -1,5 +1,7 @@
 include modules/grpc/protos/Makefile.am
 
+include modules/grpc/credentials/Makefile.am
+
 include modules/grpc/otel/Makefile.am
 include modules/grpc/loki/Makefile.am
