+++ modules/grpc/credentials/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(GRPC_CREDENTIALS_SOURCES
+    ${PROJECT_SOURCE_DIR}/modules/grpc/credentials/grpc-credentials-builder.h
+    ${PROJECT_SOURCE_DIR}/modules/grpc/credentials/grpc-credentials-builder.hpp
+    ${PROJECT_SOURCE_DIR}/modules/grpc/credentials/grpc-credentials-builder.cpp
+    PARENT_SCOPE)
