+++ modules/grpc/credentials/Makefile.am
@@ -1,0 +1,6 @@
+grpc_credentials_sources = \
+  modules/grpc/credentials/grpc-credentials-builder.h \
+  modules/grpc/credentials/grpc-credentials-builder.hpp \
+  modules/grpc/credentials/grpc-credentials-builder.cpp
+
+EXTRA_DIST +=  modules/grpc/credentials/CMakeLists.txt
